+++ pmc_automation_tools/common/utils.py
@@ -1,4 +1,4 @@
-from datetime import datetime
+from datetime import datetime, date, timedelta, timezone
 import os
 import sys
 import json
@@ -157,3 +157,25 @@ def save_updated(in_file, obj):
             c.writerows(obj)
         else:
             raise TypeError('File name provided is not an expected type of json or csv.')
+        
+
+def plex_date_formatter(*args: datetime|int, date_offset=0):
+    """
+    Takes 'normal' date formats and converts them to a Plex web service 
+        format (ISO format)
+    Can also take a single datetime object.
+    2022, 09, 11 -> 2022-09-11T04:00:00Z
+    2022, 09, 11, 18, 45 -> 2022-09-11T22:45:00Z
+        Next day if hours fall into 20-24 period
+    2022, 09, 11, 22 -> 2022-09-12T02:00:00Z
+        date_offset arg will add days to the provided time
+        Useful when providing just a datetime object to the function
+    """
+    if isinstance(args[0], (datetime, date)):
+        x = args[0]
+    else:
+        x = datetime(*args).astimezone(datetime.now(timezone.utc).tzinfo)
+    # x += timedelta(hours=4)
+    x += timedelta(days=date_offset)
+    f_date = x.strftime('%Y-%m-%dT%H:%M:%SZ')
+    return f_date